Renames env to config
