+++ .gitignore
@@ -6,8 +6,8 @@ npm-debug.log
 Dockerrun.aws.json
 *.zip
 .cmds
-src/api/env
+src/api/config
 src/api/src/models
-src/fetch/env
+src/fetch/config
 src/fetch/src/models
 .env
