+++ src/api/.sequelizerc
@@ -1,6 +1,6 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve(__dirname, 'env/db.js'),
+  'config': path.resolve(__dirname, 'config/db.js'),
   'migrations-path': path.resolve(__dirname, 'scripts/migrations'),
 };
