+++ src/api/src/resources/statistics/index.js
@@ -1,5 +1,5 @@
 const memoize = require('memoizee');
-const config = require('../../../env');
+const config = require('../../../config');
 
 module.exports = function StatisticsResource (server, controller) {
   const getStats = memoize(() => console.log('Reading stats') || Promise.all([
