+++ src/fetch/Dockerfile
@@ -9,7 +9,7 @@ WORKDIR /opt
 COPY package.json /opt/package.json
 RUN npm install --silent
 
-COPY /env /opt/env
+COPY /config /opt/config
 COPY /test /opt/test
 COPY /src /opt/src
